Added new management_type 2 (see settings file)
This new type of management is based on the MAN=VAR & REG=ON
option (as the shelterwood), with possibility to remove from above
or below.
Differently from managemente_type 1 , the number of trees to remove
is based on the current total stem volume at cell level, and the
minimum volume it is set to be left in the stand (the new variable
Prmin m3 ha-1) . When cut and the expected thinning intensity (as %)
is read in the management file. The prescribed thinning intensity is
applied as long as the minimum of the provision is mantained.
Currently, Prmin is set internally in the code. However this could
be a species-specific parameters. Hence this option works well
when the stand is monospecific.