Fix initialization values
When long term simulations soil+vegetation are performed,
and soil C value is provided, the soil C is equally split in the
modelled soil C pools.
Once equilibrium is reached, it is possible to set the new partitioning
coefficient internally in the code. In the near future, the code will read
the value of the estimated soil C pools (or partitioning coefficient) directly
from the soil file.

A small correction has been performed on the initialization of the litter pools.
As approximation, we removed the previoulsy considered CWD pool (that was simply split
in the to the litter C pools).

Please note that currently litter initialization values are considered as given from one
forest class alone (even if initialization_forest_class_litter is call within a loop across
forest classes, values in the function of initialization are not summed up).