new management type and small changes
management_type=2 option has been activated. The amount of volume to be removed
is set according to a miminum provision value to be left in the stand (this value
is currently set in the code in the management.c file), while year of thinning and
intensity is set in the management file. Please be aware that currently only the
first value of thinning intensity is considered.

Small additional changes: now the Basal area to be removed in management_type=1
is declared as cell-variable.

Thinning-threshold is compute for each class according to they management type defined in the stand
file, i.e. C (coppice) or T (timber). The year of the classes when to start the thinning
is thus set according to the forest-management or forest structure type.