read new variables in the input  soil file
Beside the usual variables provided in the soil file,
is now possible to provide also information about soil C and CWD split
in the different pools (soil_1C,soil_2C,soil_3C,soil_4C,CWD2,CWD3,CWD4)
These variables are optional. IF not provided, they are automatically
set to 0.
If only soil_C and CWD data are provided, it is assumed this
will be used as initialization value and the other subpools roughtly estimated
(this can be used to speed up the spin up process)