Mortality removal when 1 tree in the forest class alone
then there is only one tree in a forest class, we remove it in the subroutine
growing efficiency. This avoid carbon-balance issue (as we should then correct for the
height indices inside the mortality subroutine).

When age-mortality is applied, when deadtree>livetree, only one tree each time is
removed. When live tree==1, then is removed in the growthefficiencty subroutine.